feat: create measurement-sets indexs endpoint
The endpoint returns list of available MeasurementSets with pagination support.
The test asserts its behaviour.

+++ backend/src/Service/Paginator.php
@@ -48,7 +48,7 @@ class Paginator
             'total' => $total,
             'page' => $page,
             'limit' => $limit,
-            'pages' => ceil($total / $limit),
+            'pages' => (int) ceil($total / $limit),
         ];
     }
 }
